Allow for more than 4 unique colors for CPU graph
This patch changes the GSettings key from being hard-coded as
cpu-color-n to an array with the GVariant type 'a(us)'; an array of
structures containing an unsigned int (cpu index) and a string (hex
color). Using this type as opposed to a simple 'as' allows one to
easily see which color is associated with each CPU. This becomes more
important as the number of CPUs grows.

Note that this patch does not change the issue that one can only
change the color of CPUs 0-3 in the GUI. One must use a tool such as
dcong-editor to edit CPUs 4-n.

https://bugzilla.gnome.org/show_bug.cgi?id=632188