Slightly increase width of memory map dialog from 575 to 620
Normally the memory map dialog needs to be quite wide. However, the
minimum size of the main window is 626. Making a dialog wider than
this by default would be odd. Thus, this is a large as the dialog will
be by default even though it is not very useful at this width.

Normally the memory map dialog needs to be quite wide