pretty-table: prefer GIO to get application icons
Prefer Gio::AppInfo over Wnck. We need to install file monitors in the
application desktop file directories to make sure our cache is up to
date when new applications are installed or upgraded, and we're still
running.

This ensures the icons we use are consistent with those displayed by the
rest of the system.

https://bugzilla.gnome.org/show_bug.cgi?id=683762