Use a different definition of Memory
Currently, the "Memory" column is using the writable memory when
available. While being a very good heuristics for identifying private
process memory, it's damn slow to compute as it requires the whole
memory map of the process. This is the single major culprit of the
long-standing high CPU usage problem of the Process list in System Monitor.

A much faster-to-compute approximation of private memory is the
Resident Set Size (RSS), after subtracting the shared memory. While being
an underestimation, it's still the best definition, roughly corresponding to
the memory which would be freed by killing the process. This is the same value
used by other popular system monitoring tools, like KSysGuard.

This commit changes the definition of the Memory column from "writable" to
"RSS - shared". Writable memory is no longer computed for every single
process at every update of the list. Instead, it is just shown in the
property dialog of the process.

https://bugzilla.gnome.org/show_bug.cgi?id=524830