Blacklist a few executable names in gio app table
A few applications install .desktop files with an Exec= line like:

  Exec=sh -c "hexchat --existing --url %u" || exec hexchat

which makes GIO believe there's an "Hexchat" app with executable "sh".
This is problematic when we look for an icon to associate to any process
sh, because GIO gives us the Hexchat icon in this case.

Fix this by blacklisting "sh" (and "env" which seems to have the same
problem) when building the GIO application table.