load-graph.cpp: Draw the background explicitly
The gdk_window_set_background_pattern() function is a little dubious
to use unless you are writing GTK+ code since the background may be
overridden by gtkwidget.c whenever the style of the widget changes.

So instead of relying on this function, just paint the cached pattern
on every expose.

This fixes

     https://bugzilla.redhat.com/show_bug.cgi?id=1015507

in which the background grid would disappear when the
gnome-system-monitor window was unfocused, then focused.

https://bugzilla.gnome.org/show_bug.cgi?id=720560