Add more keywords to desktop file
From our review on OSNews:

Search only succeeds if you know the right keyword or characters with
which to search. For example, what if you're used to Windows and want to
watch your computer's performance? You might enter "task manager". That
won't retrieve the Linux System Monitor. Nor will "performance manager"
or "performance monitor"(*). "System" or "monitor" or some starting
substring of these two keywords is what you need to enter. The search
box is not intelligent like a Google or Bing search.

http://www.osnews.com/story/27745/My_Trip_to_GNOME_a_3_10_Review