Use real widget types, eliminate downcasting
Right now for some mysterious reason GSM's GtkBuilder code
dumbs down all widgets to GtkWidget*.
This leads to extensive downcasting and code that is
heavy-on-paper - it is not possible to determine widget type
by just looking at the GsmApplication class definition.

With this patch widgets reclaim their real types.
Upcasting is provided for where it is necessary to use widgets
as GtkWidget*.