dialog: use short process name instead of full
When there is a long process name, the name of the
process in the kill dialog is show in full. This is a problem
because long process names (such as google chrome) cause the dialog
to extend all the way across the room.

This patch fixes that by just displaying the short process name
without the command line parameters.

Signed-off-by: Ognjen Galic <smclt30p@gmail.com>

https://bugzilla.gnome.org/show_bug.cgi?id=788066