Avoid resizing the CPU usage labels when their text changes
As the number of digits in the CPU usage labels vary, their width was
varying too. When the window was narrow enough to be around the
natural width of the graphs, this could cause the width of all the
graphs on the Resources tab to resize.

Set a desired width of 6 characters on these labels. This is hopefully
enough for all values ("100.0%" is probably the longest), preventing
the size changing with the value, but it may vary depending on font.

Fixes #149, assuming the other regularly changing labels are not the
limiting factor in the width of the Resources tab.