Added new functions with Linux-only code
I have now moved my code from libgtop into GNOME System Monitor. This code, and the code for executing taskset, are now in new functions with code that is wrapped in #ifdef statements so it only gets compiled on Linux. This commit also includes a few lines of code to hide the Set Affinity menu items on all platforms except Linux.