Improve alignment of percentage values/labels for 10 or more CPUs
With 10+ cpu in the system, the cpu label in cpu box will align
ugly, this patch will make sure we count in cpu numbers to adjust
the cpu id lebal width.

And we will try to xalign cpu id label to the left while cpu load
label to the right.

Will produce a better alignment for system with many cpus.

this close #216