src: Format code-base using uncrustify
The whole codebase is a mix of coding styles which makes it difficult to
orient and to choose a style in which to further lead the code-base.

The format lines in source files have been removed as they've been
superseded by uncrustify.

Uncrustify configuration file is inspired by Mutter and Nautilus.