load-graph: Follow libadwaita style & combine grid and foreground color
GSM is a GNOME app and should follow the libadwaita style. Libadwaita
provides several CSS variables with colors and maps several legacy GTK3
CSS variables to them. Instead of relying on the legacy, use the new
variable names.

Until now the grid label colors and grid lines colors were kept
separate. That separation does not make sense to me and I unified them.
The only difference now is the alpha value that is changed for the color
of the grid.