src: Format code-base in GTK coding style using uncrustify
The whole codebase is a mix of coding styles which makes it difficult to
orient and to choose a style in which to further lead the code-base.

The format lines in source files have been removed as they've been
superseded by uncrustify.

Uncrustify configuration file is based on reference GNU C coding style
configuration[0] which has been updated to match the GTK C coding
style[1]. Some other adjustments to the configuration is taken from
Mutter[2] and Nautilus[3].

[0]
https://github.com/uncrustify/uncrustify/blob/master/etc/gnu-indent.cfg
[1] https://gitlab.gnome.org/GNOME/gtk/-/blob/main/docs/CODING-STYLE.md
[2]
https://gitlab.gnome.org/GNOME/mutter/-/blob/main/tools/uncrustify.cfg
[3]
https://gitlab.gnome.org/GNOME/nautilus/-/blob/master/data/uncrustify.cfg