load-graph: Fix two plotting issues
As far as I understand the code, the idea is to store the most recent
k+2 data points when the user wants to draw k data points. This is done
because the graph may be updated more frequently than data points are
generated, so to be able to draw a continuous line (i.e. one that does
not stop before the rightmost or leftmost vertical gridlines), two extra
points are stored that are drawn "off-screen" to make the plotted
line continuous.

However, there were two problems. First of all, `graph_delx` was
an integer. This could cause the oldest (visible) data point not to be
drawn on the leftmost vertical gridline as shown in the linked issue due
to the error caused by truncation.

Secondly, it seemed that the data points and the gridlines were not
drawn to the same scale. For example, when widening the window
the data points could freely travel into the future; when narrowing
it they could go way into the past. That is, the relative positions
of the grid lines and data points did not appear to be fixed
with respect to resizing the graph.

This change attempts to fix both issues.

To address the first issue, the step size on the x-axis is now stored
in a floating point type, so the error should be much smaller.

To address the second one, the x-axis step size is calculated with
the same width that is used for calculating the positions of the vertical
gridlines. Moreover note, that only one of the "extra" data points is
ignored when calculating the step size. This is done in order to be
able to show k data points in most cases.

For example, in the case of k = 4, a general case might look as follows:

   |               |
 ^   ^   ^   ^   ^   ^
 0   1   2   3   4   5

(The pipes mark the leftmost and rightmost vertical gridlines.)

Note that when a new data point is generated:

   |               |
   ^   ^   ^   ^   ^   ^
   1   2   3   4   5   6
                       (newest data point)

then k+1 data points may be shown.

Additionally, the x coordinate of the first visible point is set
to that of the rightmost vertical line, and later adjusted
to be outside the clip area to give the illusion of a continuous line.

Closes https://gitlab.gnome.org/GNOME/gnome-system-monitor/-/issues/293