procproperties: Move away from GtkTreeView
The changes made here try to follow process details dialog mockups in
https://gitlab.gnome.org/Teams/Design/app-mockups/-/blob/master/system-monitor/system-monitor-concept.png. Although the main goal is to get rid of GtkTreeView, that is why I took some liberties to keep showing all the information that is available in the current implementation with GtkTreeView. This is something that will have to be revisited when the rest of the interface is redesigned/restyle.

It also removes the modal and transient_for properties, since when the
selection is multiple, several dialogs can be opened at the same time.
Those properties prevented from doing it as expected, causing them to open
one on top of the other without being able to see the rest without closing
the last one.