openfiles: Port to GtkColumnView
Drops GsmTreeView without replacement as it is unused. The state of the
columns (size and position) in the current implementation were not saved.

Introduces OpenFilesData as a replacement for the deprecated GtkListStore
implementation.

Also, removes modal and transient-for properties since more than one dialog
can be opened with multiple selection at the same time.