memmaps: Port to GtkColumnView
Drops GsmTreeView without replacement as the state of the columns (size and
position) in the current implementation were not saved.

Introduces MemMapsInfo (MemMapsData name is already used) as a replacement
for the deprecated GtkListStore implementation.

Also, removes modal and transient-for properties because multiple selection
to open memmaps dialog is possible.