Replace GsmTreeView
Introduces three methods, one to load the state of the columns (order,
visibility, width and sort order) and two to save it, the first one saves
the order in an array of strings and the second one saves the sort order.

These last methods are split in two because they are connected to
items-changed of GListModel and changed of GtkSorter signals respectively,
thus reducing a little the complexity each time they are emitted.

Both the visibility and the width of the columns are binded in the loading
method so there is no need to save them afterwards.