Simplify `cgroups_enabled`
The original implementation had one bool to check whether it was initialized, and then another bool for the result. This adds an extra unnecessary step and makes the function not thread safe. The new version does not have user-visible mutable global state.

Additionally, the standard library has a function for determining whether a file exists, so use that over the Glib function.