Add unit tests for cgroup processing
This required adding a more testable function: `get_process_cgroup_name`. It accepts the contents of the cgroup file and returns the name of the cgroup. The existing `get_process_cgroup_info` function, which gets a pid from a `ProcInfo` and stores the cgroup name into the `ProcInfo`, is implemented in terms of this new function.

This also required changing the build set-up a little bit so that there is an internal-only library that is used by both the application and the tests.