graph: Update Interval and Number of Data Points Given Attention
Resource Graphs adjusted so that the Update Interval in Seconds now polls
data on the schedule specified. The Number of Data Points has been well
formed, but now the calculation matches the expected duration in both
the actual time elapsed and the label shown along the x axis.

As described in multiple Issues, there had been a hard count of 10 graph
update frames to each data poll, which was either misunderstood or
confusing. While this has been maintained to some degree, with larger
numbers of data points, the graph refresh frames will decrease based upon
a code comment from earlier times frames_per_unit = (width / num_points)

With most of the values being whole units, there is a little imprecision
involved, as a 1250 pixel width / 500 data points would select 2 or 3
pixels for each data polling. This does not seem to be a big deal as the
window can be scrolled through a wide variety of widths and will adjust
accordingly.

The key new feature is to make use of the millisecond precision in the
counter used for g_timeout_add(), so we divide the (seconds per update) /
(frames_per_unit) to call the data update function just about the right
frequency. The ugly side of this commit is the introduction of 3 blocks of
logic where g_timeout_add() is called instead of placing this in one area.

A few small things fixed. Also changed the default time for graph updates
up by 10 to 1 per second, to accomodate the earlier 10x while also trying
to eliminate labels changing too quickly for them to be readable. Also a
first step towards decreasing resource use while the graphs are being
displayed. If the data refresh rate is slower than about 1/2 a second, the
labels become increasingly difficult to read. Thought about adding a lower
limit to how often labels would redraw, but maybe in another issue.

There will be graph movement in addition to the data polling which fetches
values for display, but that is expected to enable smooth graph animation.

Adjusted maximum graph update interval from 10 to 100 seconds.

Closes #334