openfiles: Don't leak string values
Our record type was leaking every string value and always notifying,
lets clean this up.

:pointer had questionable memory management, replaced with :entry which
we acutally own, with :fd, :type, and :object being automatically
derrived from it as read-only properties.

Additionally hold the process info as :info, and since that's not an
object yet proxy it's pid, name, and icon, fields as :pid, :process,
:icon for use in lsof later.