disks: Rework disks view
We already have a property setter, there is no need to listen to notify
to discover changes.

Use a hashmap/hashset in much the same way as ‘Open Files’ to avoid so
much linear searching when refreshing.

Don't leak the volume monitor, and also ensure the settings object stays
alive.