setaffinity: Relicense as GPL-2.0-or-later
These sources, and the commits that produced them, are relicensed as
GPL-2.0-or-later.

Óscar Fernández Díaz:
I consent to the relicense to GPL-2.0-or-later.

Robert Roth:
I consent to the relicense to GPL-2.0-or-later.

Ondřej Míchal:
Thanks for trying to clear up the licesing situation. I consent to the
relicense to GPL-2.0-or-later.

Jacob Barkdull:
I consent to the relicense to GPL-2.0-or-later.

Automeris naranja:
✅ I consent to the relicense to GPL-2.0-or-later. ✅

David Stone:
I consent to the relicense to GPL-2.0-or-later.

Yannis Gerlach:
I consent to relicense any code i have committed for this repository to
be relicensed under any GPL.

See: https://gitlab.gnome.org/GNOME/gnome-system-monitor/-/merge_requests/214