actions: High-level interface for process actions
This module will wrap the basic action implementations with an async
interface that hides details around privilege acquisition from the
calling code, and also ensures the main thread is free to keep the UI
responsive.

Initially we are simply moving the existing privilege code in here,
adapted for it's new location, but these functions will soon disappear.